# Conflicts:
#	android/app/src/main/java/com/audiobookshelf/app/plugins/AbsAudioPlayer.kt

+++ android/app/src/main/java/com/audiobookshelf/app/plugins/AbsAudioPlayer.kt
@@ -338,10 +338,10 @@ class AbsAudioPlayer : Plugin() {
 
   @PluginMethod
   fun seek(call: PluginCall) {
-    val time:Int = call.getInt("value", 0) ?: 0 // Value in seconds
+    val time: Double = call.getDouble("value", 0.0) ?: 0.0 // Value in seconds, fractional
     Log.d(tag, "seek action to $time")
     mainHandler.post {
-      playerBackend.seekTo(time * 1000L)
+      playerBackend.seekTo((time * 1000L).toLong())
       call.resolve()
     }
   }
